updated ui

+++ src/screens/yelp-details.js
@@ -23,21 +23,24 @@ const YelpDetails = () => {
     const navigate = useNavigate();
     const currEmail = profile ? profile.email : ''
     const currUid = profile ? profile.uid : ''
-    console.log(profile)
     const followed = async () => {
-        const currUserBookmarks = await bookmarkService.findUserBookmarks(currEmail)
-        return currUserBookmarks.includes(businessId)
+        if (profile) {
+            const currUserBookmarks = await bookmarkService.findUserBookmarks(currEmail)
+            return currUserBookmarks.includes(businessId)
+        }
     }
     const getReviews = async () => {
         const currBusinessReviews = await reviewService.findBusinessReviews(businessId)
         setReviews(currBusinessReviews)
     }
     const handleBookmark = async () => {
-        const currUserBookmarks = await bookmarkService.findUserBookmarks(currEmail)
-        if (!followed) {
-            await bookmarkService.addBookmark(currUid, businessId, currUserBookmarks)
-        } else {
-            await bookmarkService.deleteBookmark(currUid, businessId, currUserBookmarks)
+        if (profile) {
+            const currUserBookmarks = await bookmarkService.findUserBookmarks(currEmail)
+            if (!followed) {
+                await bookmarkService.addBookmark(currUid, businessId, currUserBookmarks)
+            } else {
+                await bookmarkService.deleteBookmark(currUid, businessId, currUserBookmarks)
+            }
         }
     }
     const handlePostReview = async () => {
@@ -56,13 +59,10 @@ const YelpDetails = () => {
         getReviews();
     }, []);
     const reviewNumber = reviews.length || 0;
+    const totalRatingsArr = reviews.map(review => review.userReviewRating ? review.userReviewRating : 0)
+    const totalRatings = totalRatingsArr.length===0 ? 0 : totalRatingsArr.reduce((acc, review) => acc + review)
+    const avgRating = reviewNumber===0 ? 0 : totalRatings/reviewNumber;
     return (
-        <div className="container row">
-            <SearchBar/>
-            <div className="col-2">
-                <Nav/>
-            </div>
-            <div className="col-10">
                 <div className="row">
                     <img
                         className="ms-2 mt-4 col-2 float-end"
@@ -71,44 +71,50 @@ const YelpDetails = () => {
                     <div className="row col-10">
                         <h1 className="ms-2 mt-2">{businessDetails.name}</h1>
                         <p>
-                            <div className="ms-2">Rating: {businessDetails.rating || ""}</div>
+                            <div className="ms-2">Yelp Rating: {businessDetails.rating || ""}</div>
+                            <div className="ms-2">Our Users Rating: {avgRating===0 ? "not available" : avgRating}</div>
                             <div className="ms-2">Price: {businessDetails.price || ""}</div>
-                            <div className="ms-2">Address: {"to be updated"}</div>
+                            <div className="ms-2">Address: {businessDetails.location ? businessDetails.location.display_address.join(", ") : "to be updated"}</div>
                             <div className="ms-2">Phone Number: {businessDetails.display_phone || ""}</div>
                             <div className="ms-2">Reviews: {reviewNumber}</div>
                         </p>
                         <SecureContent>
                             <button
                                 className="btn btn-primary ms-3 col-4"
-                                onClick={() => handleBookmark}>Add to Bookmark</button>
+                                onClick={handleBookmark}>Add to Bookmark</button>
                         </SecureContent>
                     </div>
                     <div className="container">
                         <h1 className="ms-2 mt-2">
                             Reviews
                         </h1>
-                        <div>
-                            <textarea className="form-control w-75"
+                        <SecureContent>
+                            <div className="row m-3">
+                                <textarea className="me-3 col-8"
                                       onChange={(e) =>
                                           setNewReview({...newReview,
                                               userReview: e.target.value})}/>
-                            <button onClick={() => handlePostReview}
-                                    className="btn btn-primary float-end">
-                                Post Your Review
-                            </button>
-                        </div>
+                                <button onClick={handlePostReview}
+                                        className="btn btn-primary ms-3 col-2 float-end">
+                                    Post
+                                </button>
+                            </div>
+                        </SecureContent>
                         <ul className="list-group">
                             {
                                 reviews.map && reviews.map(review =>
                                     <div className="list-group-item">
                                         <div>
-                                            {review.userReview}
+                                            Posted by User: {review.firstName + " " + review.lastName}
                                         </div>
                                         <div>
-                                            {review.userReviewRating}
+                                            Review: {review.userReview}
                                         </div>
                                         <div>
-                                            {review.userReviewDate}
+                                            Rating: {review.userReviewRating}
+                                        </div>
+                                        <div>
+                                            Review Date: {review.userReviewDate}
                                         </div>
                                     </div>
                                 )
@@ -116,8 +122,6 @@ const YelpDetails = () => {
                         </ul>
                     </div>
                 </div>
-            </div>
-        </div>
     );
 };
